+++ CHANGELOG.md
@@ -4,6 +4,18 @@ This file documents notable changes to Kibana Prometheus Exporter starting with 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) starting with version 1.14.1.
 
+## 2.0.0 - 2022-02-24
+
+### Breaking changes:
+
+* Support Kibana 8.0.0. Prior versions are no longer supported, since the state api has changed.
+* Don't output timeseries if `response_times.avg_in_millis` is missing.
+
+### Others
+
+* Update Dockerfile to Python 3.10.2 and Alpine 3.15
+
+
 ## 1.17.0 - 2021-11-08
 
 * Moved container hosting to [GitHub Packages](https://github.com/vladvasiliu/kibana-prometheus-exporter-py/pkgs/container/kibana-prometheus-exporter-py)